handle offset unit in OneWireTempSensor
