+++ src/components/blocks/OneWireTempSensor/default.vue
@@ -1,3 +1,68 @@
+<script lang="ts">
+import Component from 'vue-class-component';
+
+import BlockComponent from '../BlockComponent';
+
+import { getById } from '@/store/blocks/OneWireTempSensor/getters';
+import { persist } from '@/store/blocks/OneWireTempSensor/actions';
+
+/* eslint-disable indent */
+@Component({
+  props: {
+    id: {
+      default: '',
+      type: String,
+    },
+  },
+})
+/* eslint-enable */
+export default class OneWireTempSensor extends BlockComponent {
+  addressInput = '';
+  offsetInput = 0;
+
+  get blockData() {
+    return getById(this.$store, this.$props.id);
+  }
+
+  get settings() {
+    return this.blockData.settings;
+  }
+
+  get state() {
+    return this.blockData.state;
+  }
+
+  get loading() {
+    return !!this.blockData.isLoading;
+  }
+
+  get changed() {
+    return this.settings.address !== this.addressInput ||
+      this.settings.offset.value !== this.offsetInput;
+  }
+
+  mounted() {
+    // set default values
+    this.offsetInput = this.settings.offset.value;
+    this.addressInput = this.settings.address;
+  }
+
+  save() {
+    // update offset's value
+    this.settings.offset.value = this.offsetInput;
+
+    persist(this.$store, {
+      settings: {
+        offset: this.settings.offset,
+        address: this.addressInput,
+      },
+      id: this.blockData.id,
+      serviceId: this.blockData.serviceId,
+    });
+  }
+}
+</script>
+
 <template>
   <q-card>
     <q-card-title>OneWireTempSensor ({{ id }})</q-card-title>
@@ -32,6 +97,7 @@
               stack-label="Offset"
               placeholder="Offset of sensor"
               type="number"
+              :suffix="settings.offset.unitNotation"
             />
           </q-item-main>
         </q-item>
@@ -59,9 +125,3 @@
     </q-card-main>
   </q-card>
 </template>
-
-<script lang="ts" src="./OneWireTempSensor.ts" />
-
-<style scoped>
-
-</style>
